requirements.txt
pytest==9.0.3
langwatch-scenario==0.7.9